+++ src/channel_mode.c
@@ -1334,8 +1334,8 @@ send_mode_changes_server(struct Client *source_p, struct Channel *chptr)
   parabuf[0] = '\0';
   parptr = parabuf;
 
-  mbl = snprintf(modebuf, sizeof(modebuf), ":%s TMODE %lu %s ", source_p->id,
-                 (unsigned long)chptr->creationtime, chptr->name);
+  mbl = snprintf(modebuf, sizeof(modebuf), ":%s TMODE %ju %s ", source_p->id,
+                 chptr->creationtime, chptr->name);
 
   /* Loop the list of modes we have */
   for (unsigned int i = 0; i < mode_count; ++i)
@@ -1367,8 +1367,8 @@ send_mode_changes_server(struct Client *source_p, struct Channel *chptr)
       nc = 0;
       mc = 0;
 
-      mbl = snprintf(modebuf, sizeof(modebuf), ":%s TMODE %lu %s ", source_p->id,
-                     (unsigned long)chptr->creationtime, chptr->name);
+      mbl = snprintf(modebuf, sizeof(modebuf), ":%s TMODE %ju %s ", source_p->id,
+                     chptr->creationtime, chptr->name);
 
       pbl = 0;
       parabuf[0] = '\0';
