+++ src/user.c
@@ -233,19 +233,19 @@ introduce_client(struct Client *source_p)
       continue;
 
     if (IsCapable(server, CAPAB_SVS))
-      sendto_one(server, ":%s UID %s %u %lu %s %s %s %s %s %s :%s",
+      sendto_one(server, ":%s UID %s %u %ju %s %s %s %s %s %s :%s",
                  source_p->servptr->id,
                  source_p->name, source_p->hopcount+1,
-                 (unsigned long)source_p->tsinfo,
+                 source_p->tsinfo,
                  ubuf, source_p->username, source_p->host,
                  source_p->sockhost, source_p->id,
                  source_p->account,
                  source_p->info);
     else
-      sendto_one(server, ":%s UID %s %u %lu %s %s %s %s %s :%s",
+      sendto_one(server, ":%s UID %s %u %ju %s %s %s %s %s :%s",
                  source_p->servptr->id,
                  source_p->name, source_p->hopcount+1,
-                 (unsigned long)source_p->tsinfo,
+                 source_p->tsinfo,
                  ubuf, source_p->username, source_p->host,
                  source_p->sockhost, source_p->id, source_p->info);
